# Fixed errors in JFilterInput
Also removed one assertion from JRequest-helper-dataset.php. This assertion checked if <body>stuff</body> is returned without changes by JFilterInput. However, since the XSS protection is activated, tags like <body> are stripped from the string, regardless of JREQUEST_ALLOW_HTML set or not. In the end result, it seemed as if this assertion was just plain wrong. If it is intended to still do this check, this has to go into a new test.

git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@16589 6f6e1ebd-4c2b-0410-823f-f34bde69bce9