- getFieldType(), getRuleType and getType() methods were moved to JFormHelper. Now helper is consistence.  - Ampersands before findField(), loadFieldType() and loadRuleType() were removed (see below why).  - JFormInspector modified to reflect previous change.
Explicit reference returning from some of JForm's methods

We have 6 such methods:
 - & findField (returns XML element object)
 - & loadFieldType (returns JFormField object)
 - & loadRuleType (returns JFormRule object)
 - & findFieldsByFieldset (returns array of JXMLElement objects)
 - & findFieldsByGroup (returns array of JXMLElement objects)
 - & findGroup (returns array of JXMLElement objects)

All they are defined as "protected", so it's not possible to break something unless JForm has children.
JForm has only one child: JFormInspector (this is a class from the unit tests suite) and it's not a problem to slightly modify it to reflect changes.

I think we can delete ampersands before findField, loadFieldType and loadRuleType methods without any risk because they return objects and 
we know that an object returned by reference in any case.

As for the methods which returns an array we can leave them without any change because it brings small benefit ONLY when someone will decide modify an array 
with objects (if we do not change a copy of an array then PHP will not allocate additional memory for it).



git-svn-id: http://joomlacode.org/svn/joomla/development/branches/jform-artyom@16797 6f6e1ebd-4c2b-0410-823f-f34bde69bce9